# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Royal Karahi Inventory Management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface LoginRequest {
  username: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  admin: "admin",
  user: "user",
} as const;

export interface User {
  id: number;
  username: string;
  role: UserRole;
  createdAt: string;
  /** @nullable */
  email?: string | null;
}

export interface LoginResponse {
  token: string;
  user: User;
}

export type CreateUserRequestRole =
  (typeof CreateUserRequestRole)[keyof typeof CreateUserRequestRole];

export const CreateUserRequestRole = {
  admin: "admin",
  user: "user",
} as const;

export interface CreateUserRequest {
  username: string;
  password: string;
  role: CreateUserRequestRole;
  /** @nullable */
  email?: string | null;
}

export interface Subcategory {
  id: number;
  name: string;
  categoryId: number;
  currentStock: number;
  createdAt: string;
  lowStockThreshold: number;
  costPerUnit: number;
}

export interface Category {
  id: number;
  name: string;
  subcategories: Subcategory[];
  createdAt: string;
  unit: string;
}

export interface CreateCategoryRequest {
  name: string;
  unit: string;
}

export interface CreateSubcategoryRequest {
  name: string;
  categoryId: number;
}

export interface InventoryItem {
  id: number;
  name: string;
  categoryId: number;
  categoryName: string;
  unit: string;
  currentStock: number;
  isLowStock: boolean;
  createdAt: string;
  lowStockThreshold?: number;
  costPerUnit?: number;
}

export interface InventorySummary {
  totalItems: number;
  lowStockCount: number;
  totalCategories: number;
  totalTransactionsToday: number;
}

export type CreateTransactionRequestType =
  (typeof CreateTransactionRequestType)[keyof typeof CreateTransactionRequestType];

export const CreateTransactionRequestType = {
  IN: "IN",
  OUT: "OUT",
} as const;

export interface CreateTransactionRequest {
  subcategoryId: number;
  type: CreateTransactionRequestType;
  quantity: number;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  supplierId?: number | null;
}

export type TransactionType =
  (typeof TransactionType)[keyof typeof TransactionType];

export const TransactionType = {
  IN: "IN",
  OUT: "OUT",
} as const;

export interface Transaction {
  id: number;
  subcategoryId: number;
  type: TransactionType;
  quantity: number;
  /** @nullable */
  notes?: string | null;
  userId: number;
  createdAt: string;
  /** @nullable */
  supplierId?: number | null;
}

export type TransactionDetailType =
  (typeof TransactionDetailType)[keyof typeof TransactionDetailType];

export const TransactionDetailType = {
  IN: "IN",
  OUT: "OUT",
} as const;

export interface TransactionDetail {
  id: number;
  subcategoryId: number;
  subcategoryName: string;
  categoryName: string;
  type: TransactionDetailType;
  quantity: number;
  /** @nullable */
  notes?: string | null;
  userId: number;
  username: string;
  createdAt: string;
  /** @nullable */
  supplierId?: number | null;
  /** @nullable */
  supplierName?: string | null;
  unit: string;
}

export interface ReportSummary {
  totalIn: number;
  totalOut: number;
  totalTransactions: number;
}

export interface DailyReport {
  date: string;
  transactions: TransactionDetail[];
  summary: ReportSummary;
}

export interface RangeReport {
  from: string;
  to: string;
  transactions: TransactionDetail[];
  summary: ReportSummary;
}

export interface Supplier {
  id: number;
  name: string;
  /** @nullable */
  contactInfo?: string | null;
  createdAt: string;
}

export interface CreateSupplierRequest {
  name: string;
  /** @nullable */
  contactInfo?: string | null;
}

export interface MenuItem {
  id: number;
  name: string;
  price: number;
  createdAt: string;
}

export type UpdateUserRequestRole =
  (typeof UpdateUserRequestRole)[keyof typeof UpdateUserRequestRole];

export const UpdateUserRequestRole = {
  admin: "admin",
  user: "user",
} as const;

export interface UpdateUserRequest {
  /** @nullable */
  email?: string | null;
  /** @nullable */
  password?: string | null;
  role?: UpdateUserRequestRole;
}

export interface UpdateProfileRequest {
  /** @nullable */
  email?: string | null;
  /** @nullable */
  password?: string | null;
}

export type ListInventoryParams = {
  /**
   * Search by name
   */
  search?: string;
};

export type GetDailyReportParams = {
  /**
   * Date in YYYY-MM-DD format
   */
  date: string;
};

export type GetRangeReportParams = {
  /**
   * Start date YYYY-MM-DD
   */
  from: string;
  /**
   * End date YYYY-MM-DD
   */
  to: string;
};