# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Royal Karahi Inventory Management API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication
  - name: categories
    description: Category management
  - name: inventory
    description: Inventory items
  - name: transactions
    description: Stock transactions
  - name: reports
    description: Reports
  - name: users
    description: User management
  - name: suppliers
    description: Supplier management
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags:
        - health
      summary: Health check
      responses:
        '200':
          description: Healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthStatus'
  /auth/login:
    post:
      operationId: login
      tags:
        - auth
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/me:
    get:
      operationId: getMe
      tags:
        - auth
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /categories:
    get:
      operationId: listCategories
      tags:
        - categories
      summary: List all categories with subcategories
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'
    post:
      operationId: createCategory
      tags:
        - categories
      summary: Create a category
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCategoryRequest'
      responses:
        '201':
          description: Category created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
  /categories/{id}:
    delete:
      operationId: deleteCategory
      tags:
        - categories
      summary: Delete a category
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Deleted
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /subcategories:
    post:
      operationId: createSubcategory
      tags:
        - categories
      summary: Create a subcategory
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSubcategoryRequest'
      responses:
        '201':
          description: Subcategory created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subcategory'
  /subcategories/{id}:
    delete:
      operationId: deleteSubcategory
      tags:
        - categories
      summary: Delete a subcategory
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Deleted
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /inventory:
    get:
      operationId: listInventory
      tags:
        - inventory
      summary: List all inventory items with current stock
      security:
        - bearerAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
          description: Search by name
      responses:
        '200':
          description: Inventory items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InventoryItem'
  /inventory/low-stock:
    get:
      operationId: getLowStockItems
      tags:
        - inventory
      summary: Get items with quantity below 10
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Low stock items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InventoryItem'
  /inventory/summary:
    get:
      operationId: getInventorySummary
      tags:
        - inventory
      summary: Get inventory summary stats
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Inventory summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventorySummary'
  /transactions:
    post:
      operationId: createTransaction
      tags:
        - transactions
      summary: Log a stock IN or OUT transaction
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTransactionRequest'
      responses:
        '201':
          description: Transaction created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /transactions/recent:
    get:
      operationId: getRecentTransactions
      tags:
        - transactions
      summary: Get recent transactions
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Recent transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TransactionDetail'
  /reports/daily:
    get:
      operationId: getDailyReport
      tags:
        - reports
      summary: Get daily stock report for a given date
      security:
        - bearerAuth: []
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
          description: Date in YYYY-MM-DD format
      responses:
        '200':
          description: Daily report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DailyReport'
  /reports/range:
    get:
      operationId: getRangeReport
      tags:
        - reports
      summary: Get report for a date range
      security:
        - bearerAuth: []
      parameters:
        - name: from
          in: query
          required: true
          schema:
            type: string
          description: Start date YYYY-MM-DD
        - name: to
          in: query
          required: true
          schema:
            type: string
          description: End date YYYY-MM-DD
      responses:
        '200':
          description: Range report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RangeReport'
  /users:
    get:
      operationId: listUsers
      tags:
        - users
      summary: List all users (admin only)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Users list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
    post:
      operationId: createUser
      tags:
        - users
      summary: Create a new user (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/{id}:
    delete:
      operationId: deleteUser
      tags:
        - users
      summary: Delete a user (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Deleted
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      operationId: updateUser
      tags:
        - users
      summary: Update a user (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /suppliers:
    get:
      operationId: listSuppliers
      tags:
        - suppliers
      summary: List all suppliers
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Suppliers list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Supplier'
    post:
      operationId: createSupplier
      tags:
        - suppliers
      summary: Create a supplier
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSupplierRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Supplier'
  /auth/profile:
    put:
      operationId: updateProfile
      tags:
        - auth
      summary: Update current user profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    LoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: '#/components/schemas/User'
      required:
        - token
        - user
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        role:
          type: string
          enum:
            - admin
            - user
        createdAt:
          type: string
          format: date-time
        email:
          type:
            - string
            - 'null'
      required:
        - id
        - username
        - role
        - createdAt
    CreateUserRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        role:
          type: string
          enum:
            - admin
            - user
        email:
          type:
            - string
            - 'null'
      required:
        - username
        - password
        - role
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        subcategories:
          type: array
          items:
            $ref: '#/components/schemas/Subcategory'
        createdAt:
          type: string
          format: date-time
        unit:
          type: string
      required:
        - id
        - name
        - subcategories
        - createdAt
        - unit
    Subcategory:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        categoryId:
          type: integer
        currentStock:
          type: number
        createdAt:
          type: string
          format: date-time
        lowStockThreshold:
          type: number
        costPerUnit:
          type: number
      required:
        - id
        - name
        - categoryId
        - currentStock
        - createdAt
        - lowStockThreshold
        - costPerUnit
        - lowStockThreshold
        - costPerUnit
    CreateCategoryRequest:
      type: object
      properties:
        name:
          type: string
        unit:
          type: string
      required:
        - name
        - unit
    CreateSubcategoryRequest:
      type: object
      properties:
        name:
          type: string
        categoryId:
          type: integer
      required:
        - name
        - categoryId
    InventoryItem:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        categoryId:
          type: integer
        categoryName:
          type: string
        unit:
          type: string
        currentStock:
          type: number
        isLowStock:
          type: boolean
        createdAt:
          type: string
          format: date-time
        lowStockThreshold:
          type: number
        costPerUnit:
          type: number
      required:
        - id
        - name
        - categoryId
        - categoryName
        - currentStock
        - isLowStock
        - createdAt
        - unit
    InventorySummary:
      type: object
      properties:
        totalItems:
          type: integer
        lowStockCount:
          type: integer
        totalCategories:
          type: integer
        totalTransactionsToday:
          type: integer
      required:
        - totalItems
        - lowStockCount
        - totalCategories
        - totalTransactionsToday
    CreateTransactionRequest:
      type: object
      properties:
        subcategoryId:
          type: integer
        type:
          type: string
          enum:
            - IN
            - OUT
        quantity:
          type: number
        notes:
          type:
            - string
            - 'null'
        supplierId:
          type:
            - integer
            - 'null'
      required:
        - subcategoryId
        - type
        - quantity
    Transaction:
      type: object
      properties:
        id:
          type: integer
        subcategoryId:
          type: integer
        type:
          type: string
          enum:
            - IN
            - OUT
        quantity:
          type: number
        notes:
          type:
            - string
            - 'null'
        userId:
          type: integer
        createdAt:
          type: string
          format: date-time
        supplierId:
          type:
            - integer
            - 'null'
      required:
        - id
        - subcategoryId
        - type
        - quantity
        - userId
        - createdAt
    TransactionDetail:
      type: object
      properties:
        id:
          type: integer
        subcategoryId:
          type: integer
        subcategoryName:
          type: string
        categoryName:
          type: string
        type:
          type: string
          enum:
            - IN
            - OUT
        quantity:
          type: number
        notes:
          type:
            - string
            - 'null'
        userId:
          type: integer
        username:
          type: string
        createdAt:
          type: string
          format: date-time
        supplierId:
          type:
            - integer
            - 'null'
        supplierName:
          type:
            - string
            - 'null'
        unit:
          type: string
      required:
        - id
        - subcategoryId
        - subcategoryName
        - categoryName
        - type
        - quantity
        - userId
        - username
        - createdAt
        - unit
    DailyReport:
      type: object
      properties:
        date:
          type: string
        transactions:
          type: array
          items:
            $ref: '#/components/schemas/TransactionDetail'
        summary:
          $ref: '#/components/schemas/ReportSummary'
      required:
        - date
        - transactions
        - summary
    RangeReport:
      type: object
      properties:
        from:
          type: string
        to:
          type: string
        transactions:
          type: array
          items:
            $ref: '#/components/schemas/TransactionDetail'
        summary:
          $ref: '#/components/schemas/ReportSummary'
      required:
        - from
        - to
        - transactions
        - summary
    ReportSummary:
      type: object
      properties:
        totalIn:
          type: integer
        totalOut:
          type: integer
        totalTransactions:
          type: integer
      required:
        - totalIn
        - totalOut
        - totalTransactions
    Supplier:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        contactInfo:
          type:
            - string
            - 'null'
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - createdAt
    CreateSupplierRequest:
      type: object
      properties:
        name:
          type: string
        contactInfo:
          type:
            - string
            - 'null'
      required:
        - name
    MenuItem:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        price:
          type: number
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - price
        - createdAt
    UpdateUserRequest:
      type: object
      properties:
        email:
          type:
            - string
            - 'null'
        password:
          type:
            - string
            - 'null'
        role:
          type: string
          enum:
            - admin
            - user
    UpdateProfileRequest:
      type: object
      properties:
        email:
          type:
            - string
            - 'null'
        password:
          type:
            - string
            - 'null'